+++ player/src/main/java/com/zj/player/ut/PlayerEventController.java
@@ -29,7 +29,7 @@ public interface PlayerEventController<R extends BaseRender> {
 
     void onSeekingLoading(String path, boolean isRegulate);
 
-    void onSeekChanged(int seek, int buffered, boolean fromUser, long played, long videoSize);
+    void onSeekChanged(int seek, long buffered, boolean fromUser, long played, long videoSize);
 
     @Nullable
     R getPlayerView();
